fix(avm): stop the fuzzer spending its budget on inputs it throws away
Six changes to the generator and mutator, all aimed at the fraction of inputs
that produce nothing useful.

Setup enqueued calls were mutated as often as app logic ones. A setup call that
fails makes the transaction unprovable, and it is raised before app logic runs,
so the whole input yields no simulation result. They are generated by the same
generator as app logic calls with no bias towards succeeding, so the fuzzer spent
much of its budget on transactions it discarded. Weight them down; an unprovable
transaction is still worth reaching, just not as a main activity.

Instruction block base offsets were drawn uniformly over the whole 32 bit address
space. A relative operand reaches base_offset + operand with operands at most 16
bits wide, so almost every base offset put that sum out of range and the first
relative access halted the enqueued call. Draw them within reach, and visit the
top of memory occasionally so the addressing error path stays covered.

The same sum was computed in uint32 and wrapped, in both sanitize_address_ref and
get_max_variable_address. A wrapped address made the block's memory model record
a tag at a low address the AVM never wrote, and every later operand resolved
against that block was then based on it. Saturate instead.

populate_context_from_tx_data ran before the contract artifacts were rebuilt, so
instruction generation resolved CALL and GETCONTRACTINSTANCE targets against the
previous round's addresses, and against nothing at all on a fresh input.

An encoding that exceeded max_size returned 0, which makes libFuzzer run a zero
length input; that fails to deserialize and silently runs the default program
instead, so the execution was wasted and its coverage attributed to the wrong
input. Drop programs until it fits, and leave the input unchanged if it cannot.

Gas limits were resampled uniformly over the whole processable range on every
mutation, so a limit landing near what the program actually consumes was out of
reach. Perturb from the current value with the existing uint32 mutator, floored
at the tx overhead, below which the transaction cannot start.

Over 400 executions at a fixed seed, inputs yielding a simulation result went
from 377 to 399, and unprovable transactions from 23 to 1.

The remaining numbers move by more than the changes do at this sample size; the
share of inputs consuming their whole gas limit reads anywhere between 55% and
67% across runs. The base offset change is justified by the arithmetic rather
than by that measurement, and the mutation step size is a judgement call that
needs a long campaign to confirm.

+++ barretenberg/cpp/src/barretenberg/avm_fuzzer/fuzz_lib/memory_manager.hpp
@@ -40,7 +40,7 @@ class MemoryManager {
     void set_memory_address(bb::avm2::MemoryTag tag, uint32_t address);
     std::optional<uint16_t> get_memory_offset_16(bb::avm2::MemoryTag tag, uint32_t address_index);
     std::optional<uint8_t> get_memory_offset_8(bb::avm2::MemoryTag tag, uint32_t address_index);
-    bool is_memory_address_set(uint16_t address);
+    bool is_memory_address_set(uint32_t address);
 
     std::optional<std::pair<ResolvedAddress, bb::avm2::testing::OperandBuilder>> get_resolved_address_and_operand_8(
         ParamRef address);
